removed config package. used time stdlib
